Migrate from cc.log_output to quake.log_output
